Remove GitHub discussions services and their corresponding tests for answered, total, and unanswered discussions and consolidate them as one service with different path
