--- services/github/github-discussions-unanswered.tester.js
@@ -1,13 +1,0 @@
-import { withRegex } from '../test-validators.js'
-import { createServiceTester } from '../tester.js'
-export const t = await createServiceTester()
-
-t.create('GitHub Unanswered Discussions (repo not found)')
-  .get('/not-a-user/not-a-repo.json')
-  .expectBadge({ label: 'discussions', message: 'repo not found' })
-
-const numberSpaceUnanswered = withRegex(/^\d+ unanswered$/)
-
-t.create('GitHub Unanswered Discussions (repo having discussions)')
-  .get('/vercel/next.js.json')
-  .expectBadge({ label: 'discussions', message: numberSpaceUnanswered })
